Decided to rename the project.
After much reflection on the coherence of the project's name, I decided to make a mini change in order to remain consistent. Expect new commits to fix some of the errors caused by the name change. I'd also advise you not to update to this commit for the time being, as testing is not yet complete.